requirements.txt
fastai
uvicorn
pillow
fastapi
python-multipart